move pilot display topics to separate file
